requirements.txt.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
PyJWT==2.8.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
gunicorn==20.1.0

